+++ src/gllvm.cpp
@@ -102,7 +102,6 @@ Type objective_function<Type>::operator() ()
   if(num_corlv>0){
     nlvr=0; num_lv=0; num_lv_c=0;
     quadratic=0;
-    num_RR=0;
   }
   
   // Distance matrix calculated from the coordinates for LVs
@@ -310,7 +309,7 @@ Type objective_function<Type>::operator() ()
           newlamCor(i,j) = 1;
           // newlamCor(i,j) = exp(sigmaLV(i));
         }else if(j>i){
-          newlamCor(i,j) = lambda(j+i*p-(i*(i-1))/2-2*i-1);
+          newlamCor(i,j) = lambda(num_RR*p-num_RR*(num_RR+1)/2+j+i*p-(i*(i-1))/2-2*i-1);
         }
       }
     }
@@ -327,7 +326,9 @@ Type objective_function<Type>::operator() ()
   // Variational approximation
   if((method<1) || (method>1)){
     // add offset
-    eta += offset;
+    if(offset.rows()==n){
+      eta += offset;
+    }
     // add fixed row effects
     // if(r0f.size() == n && (random(0)==0) && xr.rows() != n && r0f.rows() == n && r0f.cols() == 1){
     //   eta += r0f.replicate(1,p);
@@ -2599,39 +2600,38 @@ Type objective_function<Type>::operator() ()
             }
             
             matrix<Type> Binv(nlvr,nlvr);
-            matrix<Type> Cinv(nlvr,nlvr);
             Type logdetC;
-            Type vBinvv;
-            matrix <Type> BiQ(nlvr,nlvr);
             matrix <Type> Id(nlvr,nlvr);
             Id.setZero();Id.diagonal().fill(1.0);
             //this implementation does not follow calculation from van der Veen et al. 2021
             //but prevents Acov^-1 via woodbury matrix identity
             //see https://math.stackexchange.com/questions/17776/inverse-of-the-sum-of-matrices
+            //uses the identity (2D + A^-1) = A - 2A(I+2DA)^-1DA
+            matrix<Type>B(nlvr,nlvr);
             for (int i=0; i<n; i++) {
               Acov = A(i)*A(i).transpose();
               if(random(2)>0 && (num_lv_c+num_RR)>0)Acov += Ab_lvcov(i);
               for (int j=0; j<p;j++){
-                Cinv.setZero();
-                Binv.setZero();
-                BiQ.setZero();
-                Cinv = (Id - 2*sign*Acov*D(j)).inverse();
-                Binv = Acov+2*sign*Cinv*Acov*D(j)*Acov;
-                BiQ = Id+2*sign*D(j)*Cinv*Acov;//Q*Binv
-                
+                B = Id-sign*2*D(j)*Acov;
+                Eigen::PartialPivLU<Eigen::Matrix<Type, Eigen::Dynamic, Eigen::Dynamic>> lu(B);
+                Eigen::Matrix<Type, Eigen::Dynamic, Eigen::Dynamic> Binv = lu.inverse();
+                Type logdetC = -lu.matrixLU().diagonal().array().log().sum();
                 //the calculation generally prevents having to explicitly invert A*A^t, or having to invert A(i).
-                vBinvv = (2*newlam.col(j).transpose()*Cinv*Acov*D(j)*(sign*Acov*newlam.col(j)-2*u.row(i).transpose())+2*sign*u.row(i)*D(j)*Cinv*u.row(i).transpose()).value();
+                Type vBinvv = (2*sign*newlam.col(j).transpose()*Acov*Binv*D(j)*Acov*newlam.col(j)-4*u.row(i)*Binv*D(j)*Acov*newlam.col(j) +2*sign*u.row(i)*Binv*D(j)*u.row(i).transpose()).value();
                 
                 //extra cQ contribution for  XB,e cross term in concurrent model
                 if((random(2)<1) && (num_lv_c>0)){
-                  vBinvv += (-4*x_lv.row(i)*b_lv2*D(j)*Binv*newlam.col(j)+4*sign*u.row(i)*BiQ*D(j)*(x_lv.row(i)*b_lv2).transpose()+4*x_lv.row(i)*b_lv2*D(j)*Binv*D(j)*(x_lv.row(i)*b_lv2).transpose()).value();
-                  cQ(i,j) -= sign*2*u.row(i)*D(j)*(x_lv.row(i)*b_lv2).transpose();
+                  vBinvv += (-4*x_lv.row(i)*b_lv2*Binv*D(j)*Acov*newlam.col(j)+2*sign*x_lv.row(i)*b_lv2*Binv*D(j)*u.row(i).transpose()+2*sign*u.row(i)*Binv*D(j)*(x_lv.row(i)*b_lv2).transpose()+2*sign*x_lv.row(i)*b_lv2*Binv*D(j)*(x_lv.row(i)*b_lv2).transpose()).value();
+                  // get rid of extra terms that will occur due to the use of eta + cQ in the likelihood  
+                  cQ(i,j) -= (sign*2*u.row(i)*D(j)*(x_lv.row(i)*b_lv2).transpose()+sign*x_lv.row(i)*b_lv2*D(j)*(x_lv.row(i)*b_lv2).transpose()).value();
                 }
                 
                 //-logdetA + logdetB = logdetQ + logdetB = logdetC = det(QB^-1)
-                // partialPivLU because BiQ is asymmetric, and it ensures that it is invertible.
-                logdetC = BiQ.partialPivLu().matrixLU().diagonal().array().log().sum();
-                cQ(i,j) += 0.5*(vBinvv+logdetC) - sign*(D(j)*Acov).trace() - sign*(u.row(i)*D(j)*u.row(i).transpose()).sum();
+                // partialPivLU because B is asymmetric, and it ensures that it is invertible.
+                logdetC = Binv.partialPivLu().matrixLU().diagonal().array().log().sum();
+                cQ(i,j) += 0.5*(vBinvv+logdetC);
+                // get rid of extra terms that will occur due to the use of eta + cQ in the likelihood
+                cQ(i,j) -=  sign*(D(j)*Acov).trace() + sign*(u.row(i)*D(j)*u.row(i).transpose()).sum();
               }
             }
           }
@@ -2700,19 +2700,19 @@ Type objective_function<Type>::operator() ()
     } else if((family == 2) && (method<1)) {//binomial VA
       if (extra(0) == 0) { //logit
         for (int j=0; j<p;j++){
-        for (int i=0; i<n; i++) {
+          for (int i=0; i<n; i++) {
             // Type a = 0.5*sqrt(squeeze(eta(i,j)*eta(i,j) + 2*cQ(i,j)));//bound it because derivative logcosh = tanh(10) = 1 flattens
             // Type a = sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
-          // Type softplus_neg_a = CppAD::CondExpGt(a, Type(15), exp(-a), log1p(exp(-a)));
-          
+            // Type softplus_neg_a = CppAD::CondExpGt(a, Type(15), exp(-a), log1p(exp(-a)));
+            
             //Type b = CppAD::CondExpGt(a, 10, a/8-log(2.0), gllvmutils::logcosh(0.5*sqrt(a)));
             // Type b = CppAD::CondExpGt(a, 10, 10, gllvmutils::logcosh(0.5*sqrt(squeeze(eta(i,j)*eta(i,j) + 2*cQ(i,j)))));
-          // nll -= (y(i,j)-Ntrials(j)/2)*eta(i,j) - Ntrials(j)*(0.5*a+softplus_neg_a);//logspace_add(Type(0),-a));//gllvmutils::log1plus(exp(-a)));//log(invlogit(a)));//Ntrials(j)*gllvmutils::logcosh(a);//-0.5*tanh(0.5)*(eta(i,j)*eta(i,j)+2*cQ(i,j))+0.5*tanh(a)*(eta(i,j)*eta(i,j)+2*cQ(i,j));
-          Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
-          nll -= (y(i,j)-0.5)*eta(i,j) - logspace_add(wij, -wij);
-            
+            // nll -= (y(i,j)-Ntrials(j)/2)*eta(i,j) - Ntrials(j)*(0.5*a+softplus_neg_a);//logspace_add(Type(0),-a));//gllvmutils::log1plus(exp(-a)));//log(invlogit(a)));//Ntrials(j)*gllvmutils::logcosh(a);//-0.5*tanh(0.5)*(eta(i,j)*eta(i,j)+2*cQ(i,j))+0.5*tanh(a)*(eta(i,j)*eta(i,j)+2*cQ(i,j));
+            Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
+            nll -= (y(i,j)-Ntrials(j)*0.5)*eta(i,j) - Ntrials(j)*logspace_add(wij, -wij);
+
             if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
-             nll -= lgamma(Ntrials(j)+1.) - lgamma(y(i,j)+1.) - lgamma(Ntrials(j)-y(i,j)+1.);//norm.const.
+              nll -= lgamma(Ntrials(j)+1.) - lgamma(y(i,j)+1.) - lgamma(Ntrials(j)-y(i,j)+1.);//norm.const.
             }
           }
           // nll += n*Ntrials(j)*log(2.0);
@@ -3460,8 +3460,153 @@ Type objective_function<Type>::operator() ()
           }
         }
       }
+    }else if(family==13){ 
+      iphi = iphi/(1+iphi);
+      Type pVA;
+      if(method == 0 && extra(0)<1){
+      for (int j=0; j<p;j++){
+        for (int i=0; i<n; i++) {
+          if(!gllvmutils::isNA(y(i,j))){
+            if(y(i,j)>0){
+              nll -= log(1-iphi(j));
+              Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
+              nll -= (y(i,j)-Ntrials(j)*0.5)*eta(i,j) - Ntrials(j)*logspace_add(wij, -wij);
+              
+              if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
+                nll -= lgamma(Ntrials(j)+1.) - lgamma(y(i,j)+1.) - lgamma(Ntrials(j)-y(i,j)+1.);//norm.const.
+              }
+            }else{
+              Type LL = 0;
+              Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
+              LL += (-Ntrials(j)*0.5)*eta(i,j) - Ntrials(j)*logspace_add(wij, -wij);
+              
+              if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
+                LL += lgamma(Ntrials(j)+1.) - lgamma(Ntrials(j)+1.);//norm.const.
+              }
+              
+              pVA = exp(log(-iphi(j)+1)+LL-log((1-iphi(j))*exp(LL)+iphi(j)));
+              pVA = Type(CppAD::CondExpEq(pVA, Type(1), pVA-Type(1e-12), pVA));//check if pVA is on the boundary
+              pVA = Type(CppAD::CondExpEq(pVA, Type(0), pVA+Type(1e-12), pVA));//check if pVA is on the boundary
+              nll -= log(iphi(j))-log(1-pVA);
+            }
+          }
+        }
+      }
+      }else if(method == 0 && extra(0)>0){
+      for (int j=0; j<p;j++){
+        for (int i=0; i<n; i++) {
+          mu(i,j) = pnorm(Type(eta(i,j)),Type(0),Type(1));
+          mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(1), mu(i,j)-Type(1e-12), mu(i,j)));//check if on the boundary
+          mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(0), mu(i,j)+Type(1e-12), mu(i,j)));//check if on the boundary
+          
+          if(!gllvmutils::isNA(y(i,j))){
+            if(y(i,j)>0){
+              nll -= log(1-iphi(j));
+              nll -= y(i,j)*log(mu(i,j))+log(1-mu(i,j))*(Ntrials(j)-y(i,j));
+              nll += cQ(i,j)*Ntrials(j);
+              
+              if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
+                nll -= lgamma(Ntrials(j)+1.) - lgamma(y(i,j)+1.) - lgamma(Ntrials(j)-y(i,j)+1.);//norm.const.
+              }
+            }else{
+              Type LL = 0;
+              LL += log(1-mu(i,j))*Ntrials(j);
+              LL -= cQ(i,j)*Ntrials(j);
+              
+              if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
+                LL += lgamma(Ntrials(j)+1.) - lgamma(Ntrials(j)+1.);//norm.const.
+              }
+              
+              pVA = exp(log(-iphi(j)+1)+LL-log((1-iphi(j))*exp(LL)+iphi(j)));
+              pVA = Type(CppAD::CondExpEq(pVA, Type(1), pVA-Type(1e-12), pVA));//check if pVA is on the boundary
+              pVA = Type(CppAD::CondExpEq(pVA, Type(0), pVA+Type(1e-12), pVA));//check if pVA is on the boundary
+              nll -= log(iphi(j))-log(1-pVA);
+            }
+          }
+        }
+      }
     }
-    
+    }else if(family==14){ 
+      iphi = exp(lg_phi)/(1+exp(lg_phi) + exp(lg_phiZINB));
+      vector<Type> iphi2 = exp(lg_phiZINB)/(1+exp(lg_phi) + exp(lg_phiZINB));
+      vector<Type> iphi3 = iphi+iphi2;
+      Type pVA;
+      Type pVA2;
+      if(method == 0 && extra(0)<1){
+        for (int j=0; j<p;j++){
+          for (int i=0; i<n; i++) {
+            if(!gllvmutils::isNA(y(i,j))){
+              if(y(i,j)>0 && y(i,j)< Ntrials(j)){
+                nll -= log(1-iphi3(j));
+                Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
+                nll -= (y(i,j)-Ntrials(j)*0.5)*eta(i,j) - Ntrials(j)*logspace_add(wij, -wij);
+                
+                if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
+                  nll -= lgamma(Ntrials(j)+1.) - lgamma(y(i,j)+1.) - lgamma(Ntrials(j)-y(i,j)+1.);//norm.const.
+                }
+              }else if(y(i,j)==0){
+                Type LL = 0;
+                Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
+                LL += (-Ntrials(j)*0.5)*eta(i,j) - Ntrials(j)*logspace_add(wij, -wij);
+                
+                pVA = exp(log(1-iphi3(j))+LL-log((1-iphi3(j))*exp(LL)+iphi(j)));
+                pVA = Type(CppAD::CondExpEq(pVA, Type(1), pVA-Type(1e-12), pVA));//check if pVA is on the boundary
+                pVA = Type(CppAD::CondExpEq(pVA, Type(0), pVA+Type(1e-12), pVA));//check if pVA is on the boundary
+                nll -= log(iphi(j))-log(1-pVA);
+              }else if(y(i,j) == Ntrials(j)){
+                Type LL = 0;
+                Type wij = 0.5*sqrt(eta(i,j)*eta(i,j) + 2*cQ(i,j));
+                LL += (y(i,j)-Ntrials(j)*0.5)*eta(i,j) - Ntrials(j)*logspace_add(wij, -wij);
+                
+                pVA2 = exp(log(1-iphi3(j))+LL-log((1-iphi3(j))*exp(LL)+iphi2(j)));
+                pVA2 = Type(CppAD::CondExpEq(pVA2, Type(1), pVA2-Type(1e-12), pVA2));//check if pVA is on the boundary
+                pVA2 = Type(CppAD::CondExpEq(pVA2, Type(0), pVA2+Type(1e-12), pVA2));//check if pVA is on the boundary
+                nll -= log(iphi2(j))-log(1-pVA2);
+              }
+            }
+          }
+        }
+      }else if(method == 0 && extra(0)>0){
+        for (int j=0; j<p;j++){
+          for (int i=0; i<n; i++) {
+            mu(i,j) = pnorm(Type(eta(i,j)),Type(0),Type(1));
+            mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(1), mu(i,j)-Type(1e-12), mu(i,j)));//check if on the boundary
+            mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(0), mu(i,j)+Type(1e-12), mu(i,j)));//check if on the boundary
+            
+            if(!gllvmutils::isNA(y(i,j))){
+              if(y(i,j)>0 && y(i,j)< Ntrials(j)){
+                nll -= log(1-iphi3(j));
+                nll -= y(i,j)*log(mu(i,j))+log(1-mu(i,j))*(Ntrials(j)-y(i,j));
+                nll += cQ(i,j)*Ntrials(j);
+
+                if(Ntrials(j)>1 && (Ntrials(j)>y(i,j))){
+                  nll -= lgamma(Ntrials(j)+1.) - lgamma(y(i,j)+1.) - lgamma(Ntrials(j)-y(i,j)+1.);//norm.const.
+                }
+              }else if(y(i,j)==0){
+                Type LL = 0;
+                LL += log(1-mu(i,j))*Ntrials(j);
+                LL -= cQ(i,j)*Ntrials(j);
+                
+                
+                pVA = exp(log(1-iphi3(j))+LL-log((1-iphi3(j))*exp(LL)+iphi(j)));
+                pVA = Type(CppAD::CondExpEq(pVA, Type(1), pVA-Type(1e-12), pVA));//check if pVA is on the boundary
+                pVA = Type(CppAD::CondExpEq(pVA, Type(0), pVA+Type(1e-12), pVA));//check if pVA is on the boundary
+                nll -= log(iphi(j))-log(1-pVA);
+              }else if(y(i,j) == Ntrials(j)){
+                Type LL = 0;
+                LL += y(i,j)*log(mu(i,j))+log(1-mu(i,j))*(Ntrials(j)-y(i,j));
+                LL -= cQ(i,j)*Ntrials(j);
+                
+                pVA2 = exp(log(1-iphi3(j))+LL-log((1-iphi3(j))*exp(LL)+iphi2(j)));
+                pVA2 = Type(CppAD::CondExpEq(pVA2, Type(1), pVA2-Type(1e-12), pVA2));//check if pVA is on the boundary
+                pVA2 = Type(CppAD::CondExpEq(pVA2, Type(0), pVA2+Type(1e-12), pVA2));//check if pVA is on the boundary
+                nll -= log(iphi2(j))-log(1-pVA2);
+              }
+            }
+          }
+        }
+      }
+    }
     
   }else{
     using namespace density;
@@ -3542,7 +3687,9 @@ Type objective_function<Type>::operator() ()
     // Laplace approximation
     
     // add offset to lin. predictor 
-    eta += offset;
+    if(offset.rows()==n){
+      eta += offset;
+    }
     // if(r0f.size() == n && (random(0)==0) && xr.rows() != n && r0f.rows() == n && r0f.cols() == 1){
     //   eta += r0f.replicate(1,p);
     if(xr.rows()==n){
@@ -4142,6 +4289,45 @@ Type objective_function<Type>::operator() ()
               nll -= log(1-iphi(j)) + dnbinom_robust(y(i,j), eta(i,j), 2*eta(i,j) - lg_phiZINB(j), 1);
             }else{
               nll -= log(iphi(j) + (Type(1)-iphi(j))*dnbinom_robust(y(i,j), eta(i,j), 2*eta(i,j) - lg_phiZINB(j), 0)); 
+            }
+          }
+        }
+      }
+    }else if(family= 13){
+      iphi=iphi/(1+iphi);
+      for (int j=0; j<p;j++){
+        for (int i=0; i<n; i++) {
+          if(extra(0)<1) {mu(i,j) = mu(i,j)/(mu(i,j)+1);
+          } else {mu(i,j) = pnorm(eta(i,j));}
+          mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(1), mu(i,j)-Type(1e-12), mu(i,j)));//check if on the boundary
+          mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(0), mu(i,j)+Type(1e-12), mu(i,j)));//check if on the boundary
+          if(!gllvmutils::isNA(y(i,j))){
+            if(y(i,j)>0){
+              nll -= log(1-iphi(j)) + dbinom(y(i,j), Type(Ntrials(j)), mu(i,j), 1);
+            }else{
+              nll -= log(iphi(j) + (Type(1)-iphi(j))*dbinom(y(i,j), Type(Ntrials(j)), mu(i,j), 0)); 
+            }
+          }
+        }
+      }
+    }else if(family= 14){
+      iphi = exp(lg_phi)/(1+exp(lg_phi) + exp(lg_phiZINB));
+      vector<Type> iphi2 = exp(lg_phiZINB)/(1+exp(lg_phi) + exp(lg_phiZINB));
+      vector<Type> iphi3 = iphi+iphi2;
+      for (int j=0; j<p;j++){
+        for (int i=0; i<n; i++) {
+          if(extra(0)<1) {mu(i,j) = mu(i,j)/(mu(i,j)+1);
+          } else {mu(i,j) = pnorm(eta(i,j));}
+          mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(1), mu(i,j)-Type(1e-12), mu(i,j)));//check if on the boundary
+          mu(i,j) = Type(CppAD::CondExpEq(mu(i,j), Type(0), mu(i,j)+Type(1e-12), mu(i,j)));//check if on the boundary
+          if(!gllvmutils::isNA(y(i,j))){
+            if(y(i,j)>0 && y(i,j) < Ntrials(j)){
+              nll -= log(1-iphi3(j)) + dbinom(y(i,j), Type(Ntrials(j)), mu(i,j), 1);
+            }else if(y(i,j)==0){
+              nll -= log(iphi(j) + (Type(1)-iphi3(j))*dbinom(y(i,j), Type(Ntrials(j)), mu(i,j), 0)); 
+            }else if(y(i,j) == Ntrials(j)){
+              nll -= log(iphi2(j) + (Type(1)-iphi3(j))*dbinom(y(i,j), Type(Ntrials(j)), mu(i,j), 0)); 
+              
             }
           }
         }